Crate test_allocator

This create provides a simple bump allocator useful for tests of determinism. Because it never reuses memory, that’s just about all it’s useful for.

The allocator provides deterministic result addresses and retired conditional branch counts in circumstances other allocators don’t. When running several tests of determinism forked from the same process, allocations in between test runs will alter the allocator state. This normally results in different allocation addresses and codepaths taken.

Determinism is provided in these circumstances by calling the skip_to_offset method with a high enough value. If successful, this sets the allocator state to a deterministic point, after which both addresses and branch counts should be consistent. Racing threads that allocate will break this guarantee: it is not known which thread will succeed and which will need to retry, getting a different address and executing additional branches.

Under the hood, the allocator is backed by a huge mmaped region at a known address. Linux overcommit behavior ensures the system does not run out of memory. Allocations are lock-free and bump downwards.

skip_to_offset simply sets the pointer for the next allocation to the requested distance from the top of the slab, failing if that region is already used.

Structs

Global

OvercommitBumpAlloc

A bump allocator backed by an overcommitted mmap(2) region that never deallocates.

Enums

SkipError

Constants

GLOBAL_MAX_OFFSET

The maximum argument to GLOBAL.skip_to_offset

GLOBAL_SLAB_TOP

Use this constant to calculate expected addresses in conjunction with skip_to_offset.

Statics

GLOBAL